Migrate Phonetic method to use Data structure Remove database.rs of Phonetic module
