+++ src/phonetic/mod.rs
@@ -1,4 +1,3 @@
-mod database;
 pub(crate) mod method;
 mod regex;
 mod suggestion;
